requirements.microservices.txt
# HTBase Microservices Requirements

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Celery and Redis
celery[redis]>=5.3.0
redis>=5.0.0
kombu>=5.3.0
flower>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Cloud Storage
google-cloud-storage>=2.14.0

# Content extraction
readability-lxml>=0.8.1
lxml>=5.0.0

# Utilities
python-dotenv>=1.0.0

# Monitoring (optional)
prometheus-client>=0.19.0
